less verbose logging, consistent log format
